@GuillaumeGomez 2020-04-13
Thanks a lot! I'll make a small release later today.